requirements.txt
click==8.1.3
Flask==2.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.1
numpy==1.21.6
pandas==1.1.5
py4j==0.10.7
pymongo==4.1.1
pyspark==2.4.7
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
typing-extensions==4.2.0
Werkzeug==2.1.2
zipp==3.8.0
